Updated the AdminAddProduct to use the storage bucket
